Add EchoRequest and EchoResponse (#4)
This implements sending an EchoRequest and receiving an EchoResponse as
defined by the LIFX LAN protocol. I have found that GetPower and
GetHardwareVersion are sometimes spotty and I'd like a method that
allows me to make sure my devices are online without actually requesting
information from the device.

If a payload of 64 \x00 bytes is sent to a device, the device treats it
as an empty payload. It then returns a default valued payload, which is
not all \x00 values, and is therefore unexpected.